requirements.txt
fastapi
uvicorn
python-multipart
jinja2
pymupdf
requests
groq
python-dotenv
